fourth/cloud/exo/requirements.txt
attrs==22.1.0
boto3==1.24.84
botocore==1.27.84
certifi==2022.9.24
charset-normalizer==2.1.1
cs==3.0.0
exoscale==0.7.1
idna==3.4
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
requests-exoscale-auth==1.1.2
s3transfer==0.6.0
six==1.16.0
toml==0.10.2
urllib3==1.26.12
